The 'observation' parameter in the double probability(observation) function has been changed to long to be consistent with the type of the draw() function.

+++ trunk/dsol-core/src/main/java/nl/tudelft/simulation/jstats/distributions/DistDiscreteConstant.java
@@ -43,7 +43,7 @@ public class DistDiscreteConstant extends DistDiscrete
 
     /** {@inheritDoc} */
     @Override
-    public double probability(final int observation)
+    public double probability(final long observation)
     {
         if (observation == this.constant)
         {
